split mixed_integer_rotation_constraint_test
